+++ src/main/java/com/example/QuickBid/model/User.java
@@ -1,8 +1,7 @@
 package com.example.QuickBid.model;
 
 import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
@@ -15,7 +14,7 @@ import java.time.LocalDateTime;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long UserId;
+    private int UserId;
 
     @Column(nullable = false, length = 50)
     private String fullname;
@@ -37,7 +36,6 @@ public class User {
 
     @Column(name = "createdAt")
     private LocalDateTime createdAt;
-
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "whoApproved", nullable = true)
