+++ src/main/java/com/example/QuickBid/repository/AdminRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AdminRepository extends JpaRepository<Admin, Long> {
+public interface AdminRepository extends JpaRepository<Admin, Integer> {
 
     Optional<Admin> findByAdminUsername(String adminUsername);
